change mcr-complete.js to mcr-complete.ts
