+++ src/main/ts/index.js
@@ -20,7 +20,7 @@ import Hand from './jan/hand';
 import {Janpai} from './jan/janpai';
 import {JANPAI_ID} from './jan/janpai-id';
 import JanUtil from './jan/jan-util';
-import McrComplete from './jan/mcr-complete';
+import {McrComplete} from './jan/mcr-complete';
 import McrUtil from './jan/mcr-util';
 import {MCR_YAKU} from './jan/mcr-yaku';
 import Mentsu from './jan/mentsu';
