+++ src/main/ts/jan/mcr-complete.ts
@@ -1,0 +1,47 @@
+/*
+Copyright 2021 masaue
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+ */
+import {McrYaku} from './mcr-yaku';
+
+export type McrCompleteObject = {yakuList: McrYaku[]};
+
+export class McrComplete {
+  private _point;
+
+  constructor(private _yakuList: McrYaku[]) {
+    this._point = this._totalPoint(_yakuList);
+  }
+
+  toObject(): McrCompleteObject {
+    return {yakuList: this._yakuList};
+  }
+
+  get point() {
+    return this._point;
+  }
+
+  get yakuList() {
+    return this._yakuList;
+  }
+
+  private _totalPoint(yakuList: McrYaku[]) {
+    if (yakuList.length === 0) {
+      return 0;
+    }
+    const yakuPointList: number[] = yakuList.map((y) => { return y.point; })
+
+    return yakuPointList.reduce((p, c) => { return p + c; } );
+  }
+}
